DateGeneration deserved it's own class at this point
